+++ cogito/get.go
@@ -33,7 +33,6 @@ func Get(log hclog.Logger, input []byte, out io.Writer, args []string) error {
 	if err != nil {
 		return err
 	}
-	log.SetLevel(hclog.LevelFromString(request.Source.LogLevel))
 	log.Debug("parsed get request",
 		"source", request.Source,
 		"version", request.Version,